@@ -32,53 +32,79 @@
     "e2e": "ng e2e",
     "copyhtml": "copy 'src/app/cron-editor/*.html' './components/cron-editor/'",
     "copycss": "copy 'src/app/cron-editor/*.css' './components/cron-editor/'",
-    "pack": "ng-packagr -p package.json & cd dist & npm pack"
+    "pack": "ng-packagr -p package.json",
+    "publish": "ng-packagr -p package.json && cd dist && npm publish"
   },
   "dependencies": {
-    "@angular/animations": "^6.1.10",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.6",
-    "@angular/compiler": "^6.1.6",
-    "@angular/core": "^6.1.6",
-    "@angular/forms": "^6.1.6",
-    "@angular/http": "^6.1.6",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^6.1.6",
-    "@angular/platform-browser-dynamic": "^6.1.6",
-    "@angular/router": "^6.1.6",
-    "bootstrap": "4.0.0-beta.2",
-    "core-js": "^2.5.1",
-    "jquery": "1.9.1 - 3",
-    "popper.js": "^1.12.3",
-    "rxjs": "6.2.2",
-    "zone.js": "^0.8.17"
+    "tslib": "^1.9.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
-    "@angular/cli": "^6.2.5",
-    "@angular/compiler-cli": "^6.1.6",
-    "@angular/language-service": "^6.1.6",
-    "@types/jasmine": "^2.8.9",
-    "@types/jasminewd2": "^2.0.5",
-    "@types/node": "^8.10.36",
-    "codelyzer": "~4.0.2",
-    "jasmine-core": "~2.8.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
+    "@angular-devkit/build-angular": "~0.801.0",
+    "@angular-devkit/build-ng-packagr": "~0.6.8",
+    "@angular/animations": "^8.1.0",
+    "@angular/cdk": "^8.0.2",
+    "@angular/cli": "^8.1.0",
+    "@angular/common": "^8.1.0",
+    "@angular/compiler": "^8.1.0",
+    "@angular/compiler-cli": "8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/forms": "^8.1.0",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.2",
+    "@angular/language-service": "^8.1.0",
+    "@angular/platform-browser": "^8.1.0",
+    "@angular/platform-browser-dynamic": "^8.1.0",
+    "@angular/router": "^8.1.0",
+    "@types/jasmine": "^3.3.13",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/node": "^12.6.0",
+    "codelyzer": "^5.1.0",
+    "core-js": "^3.1.4",
+    "jasmine-core": "^3.4.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^3.0.6",
-    "protractor": "~5.2.2",
-    "ts-node": "~4.0.1",
-    "tslint": "~5.8.0",
-    "typescript": "2.7.2"
+    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma-jasmine": "^2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^5.5.0",
+    "protractor": "^5.4.2",
+    "rxjs": "6.5.2",
+    "ts-node": "^8.3.0",
+    "tslint": "~5.18.0",
+    "typescript": "~3.4.5",
+    "zone.js": "^0.9.1"
+  },
+  "peerDependencies": {
+    "@angular/animations": "^8.1.0",
+    "@angular/cdk": "^8.0.2",
+    "@angular/common": "^8.1.0",
+    "@angular/compiler": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@angular/forms": "^8.1.0",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^8.0.2",
+    "@angular/platform-browser": "^8.1.0",
+    "@angular/platform-browser-dynamic": "^8.1.0",
+    "@angular/router": "^8.1.0",
+    "core-js": "^3.1.4",
+    "rxjs": "6.5.2",
+    "zone.js": "^0.9.1"
   },
   "ngPackage": {
     "lib": {
-      "entryFile": "src/app/cron-editor/cron-editor.module.ts"
+      "entryFile": "index.ts"
     }
-  }
+  },
+  "main": "bundles/ngx-cron-editor.umd.js",
+  "module": "fesm5/ngx-cron-editor.js",
+  "es2015": "fesm2015/ngx-cron-editor.js",
+  "esm5": "esm5/ngx-cron-editor.js",
+  "esm2015": "esm2015/ngx-cron-editor.js",
+  "fesm5": "fesm5/ngx-cron-editor.js",
+  "fesm2015": "fesm2015/ngx-cron-editor.js",
+  "typings": "ngx-cron-editor.d.ts",
+  "metadata": "ngx-cron-editor.metadata.json",
+  "sideEffects": false
 }
