@@ -1,13 +1,13 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
+import {BrowserModule} from '@angular/platform-browser';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
 
-import { CronEditorModule } from './cron-editor/cron-editor.module';
+import {CronEditorModule} from './cron-editor/cron-editor.module';
 
-import { AppComponent } from './app.component';
+import {AppComponent} from './app.component';
 
 @NgModule({
-  imports: [BrowserModule, FormsModule, CronEditorModule],  
+  imports: [BrowserModule, FormsModule, CronEditorModule],
   declarations: [AppComponent],
   providers: [],
   bootstrap: [AppComponent]
